+++ tomni/bbox_fitting/bbox_fitting_center_by_scale/test_bbox_fitting_center_by_scale.py
@@ -5,7 +5,7 @@ import numpy as np
 from .main import crop_image_by_scale
 
 
-class TestCropByScale(TestCase):
+class TestBboxFittingCenterByScale(TestCase):
     @classmethod
     def setUp(self):