+++ tomni/img_center_crop/crop_by_dim/__init__.py
@@ -1,1 +1,0 @@
-from .main import crop_image_by_dim